fix(package): vue should be a peer dep, details follow
As a plugin, this should not hard depend on Vue, but rather specify a
compatible range in `peerDependencies`, most libraries/plugins
(including those in the Vue ecosystem) seem to be using this pattern.
A hard depend (as of now), means that it will result in multiple Vue
installs and even loads at runtime.